split out table validator
